code review refactor
