+++ src/Core/Grid/Definition/Factory/CustomerThreadGridDefinitionFactory.php
@@ -28,7 +28,7 @@ declare(strict_types=1);
 
 namespace PrestaShop\PrestaShop\Core\Grid\Definition\Factory;
 
-use PrestaShop\PrestaShop\Adapter\CustomerService\ContactTypeProvider;
+use PrestaShop\PrestaShop\Core\Domain\CustomerService\DataProvider\ContactTypeProvider;
 use PrestaShop\PrestaShop\Core\Form\ConfigurableFormChoiceProviderInterface;
 use PrestaShop\PrestaShop\Core\Form\FormChoiceProviderInterface;
 use PrestaShop\PrestaShop\Core\Grid\Action\Bulk\BulkActionCollection;
