+++ src/PrestaShopBundle/Resources/config/services/adapter/customer_service.yml
@@ -2,15 +2,15 @@ services:
   _defaults:
     public: true
 
-  prestashop.adapter.customer_service.query_handler.get_customer_thread_for_viewing_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\CustomerService\QueryHandler\GetCustomerThreadForViewingHandler'
+  prestashop.ore.domain.customer_service.query_handler.get_customer_thread_for_viewing_handler:
+    class: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\QueryHandler\GetCustomerThreadForViewingHandler'
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
     tags:
       - name: tactician.handler
         command: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\Query\GetCustomerThreadForViewing'
 
-  prestashop.adapter.customer_service.query_handler.get_customer_service_signature_handler:
+  prestashop.core.domain.customer_service.query_handler.get_customer_service_signature_handler:
     class: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\QueryHandler\GetCustomerServiceSignatureHandler'
     arguments:
       - '@prestashop.adapter.legacy.configuration'
@@ -18,8 +18,8 @@ services:
       - name: tactician.handler
         command: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\Query\GetCustomerServiceSignature'
 
-  prestashop.adapter.customer_service.command_handler.forward_customer_thread_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\CustomerService\CommandHandler\ForwardCustomerThreadHandler'
+  prestashop.core.domain.customer_service.command_handler.forward_customer_thread_handler:
+    class: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\CommandHandler\ForwardCustomerThreadHandler'
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
       - '@prestashop.adapter.legacy.configuration'
@@ -27,16 +27,16 @@ services:
       - name: tactician.handler
         command: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\Command\ForwardCustomerThreadCommand'
 
-  prestashop.adapter.customer_service.command_handler.reply_to_customer_thread_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\CustomerService\CommandHandler\ReplyToCustomerThreadHandler'
+  prestashop.core.domain.customer_service.command_handler.reply_to_customer_thread_handler:
+    class: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\CommandHandler\ReplyToCustomerThreadHandler'
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
     tags:
       - name: tactician.handler
         command: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\Command\ReplyToCustomerThreadCommand'
 
-  prestashop.adapter.customer_service.command_handler.update_customer_thread_status_handler:
-    class: 'PrestaShop\PrestaShop\Adapter\CustomerService\CommandHandler\UpdateCustomerThreadStatusHandler'
+  prestashop.core.domain.customer_service.command_handler.update_customer_thread_status_handler:
+    class: 'PrestaShop\PrestaShop\Core\Domain\CustomerService\CommandHandler\UpdateCustomerThreadStatusHandler'
     arguments:
       $connection: '@doctrine.dbal.default_connection'
       $dbPrefix: '%database_prefix%'
