+++ src/PrestaShopBundle/Resources/config/services/adapter/data_provider.yml
@@ -106,7 +106,7 @@ services:
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext().employee'
 
-  prestashop.adapter.data_provider.contact_type:
-    class: PrestaShop\PrestaShop\Adapter\CustomerService\ContactTypeProvider
+  prestashop.core.domain.customer_service.data_provider.contact_type:
+    class: PrestaShop\PrestaShop\Core\Domain\CustomerService\DataProvider\ContactTypeProvider
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext().language.id'
