+++ src/PrestaShopBundle/Resources/config/services/core/grid/grid_definition_factory.yml
@@ -342,7 +342,7 @@ services:
     class: 'PrestaShop\PrestaShop\Core\Grid\Definition\Factory\CustomerThreadGridDefinitionFactory'
     parent: 'prestashop.core.grid.definition.factory.abstract_grid_definition'
     arguments:
-      - '@prestashop.adapter.data_provider.contact_type'
+      - '@prestashop.core.domain.customer_service.data_provider.contact_type'
       - '@prestashop.adapter.form.choice_provider.shop_name_by_id'
       - '@prestashop.core.form.choice_provider.customer_thread_statuses_choice_provider'
     public: true
